+++ CHANGELOG.md
@@ -9,12 +9,14 @@ ADDED:
 - new constructor with a set from exclude values in class EnumComboBoxModel
 - new class GenericComboBoxModel with generic type value
 - new class GenericMutableComboBoxModel with generic type value
-- 
+
 CHANGED:
 
 - update of gradle-plugin dependency 'com.diffplug.spotless:spotless-plugin-gradle' to new version 6.11.0
 - update of dependency jobj-core to new major version 7
 - update of dependency gen-tree to new major version 7
+- update of dependency model-data to new major version 2
+- update of dependency swing-base-components to new major version 2
 - update of test dependency file-worker to new minor version 11.3
 - update of test dependency zip4j to new patch version 2.11.2
 - update of test dependency junit-jupiter to new minor version 5.9.1